requirements.txt
# Han.Eye SaaS Requirements
# Python 3.9+

# Flask Framework
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5

# Database
SQLAlchemy==2.0.23

# AI/ML Libraries
openai==1.6.1
anthropic==0.8.1
google-generativeai==0.3.2

# Computer Vision
opencv-python==4.8.1.78
Pillow==10.1.0
numpy==1.26.2

# HTTP Requests
requests==2.31.0

# Security
Werkzeug==3.0.1

# Utilities
python-dotenv==1.0.0

# Reinforcement Learning
stable-baselines3==2.2.1
gymnasium==0.29.1

# Development
pytest==7.4.3
pytest-cov==4.1.0

